Issue #2183: Autoscaled cluster costs - system preferences support
